Updating UI and character map.
