+++ lib/UI/character_ui.rb
@@ -1,0 +1,5 @@
+require_relative './basic_ui.rb'
+
+module CharacterUI
+	include BasicUI
+end
