+++ lib/UI/game_ui.rb
@@ -1,5 +1,9 @@
+require_relative './basic_ui.rb'
+
 # Adding UI specific for game-based utilities
 module GameUI
+  include BasicUI
+
   def start_menu_display
     system 'clear'
     puts "\n"
@@ -69,31 +73,6 @@ module GameUI
     puts Paint[character.name.upcase, :italic]
     puts "Level: #{character.level}"
     puts "HP: #{character.hp}/#{character.max_hp}".ljust(10) + "HP: #{character.mana}/#{character.max_mana}".ljust(10)
-    puts '-' * 30
-  end
-
-  def draw_character_map(map, location, x_max, y_max, x_min, y_min)
-    puts Paint['YOUR MAP', :bold]
-    puts '-' * 30
-    (y_min..y_max).reverse_each do |y|
-      (x_min..x_max).each do |x|
-        if map[[x, y]].nil?
-          print_icon(' ')
-        elsif map[[x, y]] == 'wall'
-          print_icon('.')
-        elsif location.grid_point == [x, y]
-          print_icon('C')
-        else
-          print_icon(map[[x, y]].icon)
-        end
-      end
-      new_line(2)
-    end
-    puts '-' * 30
-    puts Paint['KEY', :bold]
-    puts  'You Are Here!: ' + Paint['C', :bold]
-    puts 'Wall: ' + Paint['.'.ljust(2), :bold] + 'Hall: ' + Paint['h'.ljust(2), :bold] + 
-         'Room: ' + Paint['r', :bold]
     puts '-' * 30
   end
 end
